requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.21.0
asyncpg==0.30.0
alembic==1.15.2
httpx==0.28.1
python-dotenv==1.1.0
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
livekit
livekit-agents
livekit-api
google-genai==0.6.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-multipart==0.0.9
aiofiles==24.1.0
websockets==15.0
